+++ babel.config.js
@@ -1,0 +1,11 @@
+const presets = []
+const plugins = []
+
+if (process.env.NODE_ENV === 'test') {
+  // Convert to commonJS when running in jest
+  presets.unshift(['@babel/preset-env', {
+    modules: 'commonjs'
+  }])
+}
+
+module.exports = { presets, plugins }