+++ e2e/harness/src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="app">
     <router-link to="/" tag="button" id='home-button'> Home </router-link>
-    <button v-if='authenticated' v-on:click='logout' id='logout-button'> Logout </button>
+    <button v-if='authState.isAuthenticated' v-on:click='logout' id='logout-button'> Logout </button>
     <button v-else v-on:click='login' id='login-button'> Login </button>
     <router-link to="/protected" tag="button"> Protected </router-link>
     <router-view/>
@@ -12,27 +12,12 @@
 
 export default {
   name: 'app',
-  data: function () {
-    return { authenticated: false }
-  },
-  created () { this.isAuthenticated() },
-  watch: {
-    // Everytime the route changes, check for auth status
-    '$route': 'isAuthenticated'
-  },
   methods: {
-    async isAuthenticated () {
-      this.authenticated = await this.$auth.isAuthenticated()
-    },
     login () {
-      this.$auth.loginRedirect('/')
+      this.$auth.signInWithRedirect()
     },
     async logout () {
-      await this.$auth.logout()
-      await this.isAuthenticated()
-
-      // Navigate back to home
-      this.$router.push({ path: '/' })
+      await this.$auth.signOut()
     }
   }
 }