+++ e2e/harness/src/router/index.js
@@ -4,7 +4,8 @@ import Router from 'vue-router'
 import Protected from '@/components/Protected'
 import SessionTokenLogin from '@/components/SessionTokenLogin'
 
-import Auth from '@okta/okta-vue'
+import OktaVue, { LoginCallback } from '@okta/okta-vue'
+import { OktaAuth } from '@okta/okta-auth-js'
 
 // To perform end-to-end PKCE flow we must be configured on both ends: when the login is initiated, and on the callback
 // The login page is loaded with a query param. This will select a unique callback url
@@ -30,20 +31,20 @@ if (CONFIG.OKTA_TESTING_DISABLEHTTPSCHECK) {
   }
 }
 
+const oktaAuth = new OktaAuth(config)
+
 Vue.use(Router)
-Vue.use(Auth, config)
+Vue.use(OktaVue, { oktaAuth })
 
 const router = new Router({
   mode: 'history',
   base: '/',
   routes: [
-    { path: '/implicit/callback', component: Auth.handleCallback() },
-    { path: '/pkce/callback', component: Auth.handleCallback() },
+    { path: '/implicit/callback', component: LoginCallback },
+    { path: '/pkce/callback', component: LoginCallback },
     { path: '/protected', component: Protected, meta: { requiresAuth: true } },
     { path: '/sessionToken', component: SessionTokenLogin }
   ]
 })
-
-router.beforeEach(Vue.prototype.$auth.authRedirectGuard())
 
 export default router