+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+[authState]: https://github.com/okta/okta-auth-js#authstatemanager
+
+# 3.0.0
+
+### Breaking Changes
+
+- [#27](https://github.com/okta/okta-vue/pull/27) See [MIGRATING](MIGRATING.md) for detailed information.
+  - Replaces `Auth` service with instance of [@okta/okta-auth-js](https://github.com/okta/okta-auth-js) for `$auth`, so all configuration options and public methods are available.
+  - By default `isAuthenticated` will be true only if **both** accessToken **and** idToken are valid
+  - Replaces `Auth.handleCallback` with `LoginCallback` component
+  - Changes `@okta/okta-auth-js`, `vue` and `vue-router` as peerDependencies
+
+### Features
+
+- [#27](https://github.com/okta/okta-vue/pull/27)
+  - Injects reactive [authState][] property to the Vue components
+  - Guards secure routes not only by the `beforeRouteEnter` navigation guard, also by subscribing to [authState.isAuthenticated][authState] change after entering the routes.
+- [#31](https://github.com/okta/okta-vue/pull/31) Exposes ES module bundle from package module field
+
 # 2.1.1
 
 ### Bug Fixes